feat!: remove deprecated receivers resource (v4.0.0)
BREAKING CHANGE: the receivers resource has been removed. All callers
must migrate to customers (added in v3.11.0). Type names that were
Customer-prefixed in v3.11.0 to avoid collisions with the receivers
module are now renamed to their canonical short names:

- CustomerOwner -> Owner
- CustomerKycType -> KycType
- CustomerFraudWarning -> FraudWarning
- CustomerIndividualWithStandardKYC -> IndividualWithStandardKYC
- CreateCustomerIndividualWithStandardKYCInput
    -> CreateIndividualWithStandardKYCInput
- GetCustomerLimitIncreaseRequestsResponse
    -> GetLimitIncreaseRequestsResponse
- ...and so on for every collision name

Method signatures match what receivers exposed; only the accessor
name and field names (receiver_id -> customer_id) change. Dependent
resources (bank-accounts, virtual-accounts, wallets) now exclusively
call /v1/instances/{id}/customers/{customer_id}/... URLs and accept
customer_id in their input types.

Recommended merge timing: hold this PR open until the API actually
drops /receivers routes (planned 2026-07-03). 3.11.0 users on
/receivers keep working via the API's 301 redirect through the
deprecation window; publishing 4.0.0 before then risks orphaning
users mid-migration.

https://www.blindpay.com/changelog/2026-06-04-customers-rename
