+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@blindpay/node",
-    "version": "3.11.0",
+    "version": "4.0.0",
     "description": "Official Node.js SDK for Blindpay API - Stablecoin API for global payments",
     "keywords": [
         "blindpay",
