+++ src/index.ts
@@ -10,7 +10,6 @@ export * from "./resources/payins";
 export * from "./resources/payins/quotes";
 export * from "./resources/payouts";
 export * from "./resources/quotes";
-export * from "./resources/receivers";
 export * from "./resources/transfers";
 export * from "./resources/upload";
 export * from "./resources/virtual-accounts";
